Change default model config to G1 baseline, including FP32 gradients

+++ axlearn/common/learner.py
@@ -521,7 +521,7 @@ class AccumulatedLearner(Learner):
         )
 
         def _copy_zero(model_tree):
-            return jax.tree_map(lambda x: jnp.full_like(x, 0, dtype=jnp.bfloat16), model_tree)
+            return jax.tree_map(lambda x: jnp.full_like(x, 0, dtype=jnp.float32), model_tree)
 
         def run_microbatch(gradient_buffer, microbatch):
             gradient_buffer, forward_outputs_buffer = gradient_buffer
@@ -531,7 +531,7 @@ class AccumulatedLearner(Learner):
                 inputs=microbatch,
                 should_compute_gradients=should_compute_gradients,
             )
-            microbatch_gradients = jax.tree_map(lambda x: x.astype(jnp.bfloat16), microbatch_gradients)
+            microbatch_gradients = jax.tree_map(lambda x: x.astype(jnp.float32), microbatch_gradients)
 
             # accumulate gradients
             gradient_buffer = jax.tree_map(lambda x, y: x + y, microbatch_gradients, gradient_buffer)
